Fix #42: Dashboard Group not removed if unmapped directory user not in input set
The problem was that only users in the input set were actually having their groups updated.  Now all users in the Adobe directory, whether they match input users or not, are having their groups updated (which in this case means mapped groups are always removed, because if they are not in the input set we are not calculating their mapped groups).

I also added a lot of comments reflecting the code analysis I had to do.