Complete work to fix #90, fix #46, and fix #101.
This ended up being a major overhaul of the code that processes Adobe users who don't match users on the customer side.  The old code was overly complicated, because it was trying to carefully sequence deleting membership in groups with membership in organizations.  But in fact this is not necessary, because removing a user from an organization will remove that user from all the groups in the organization.  So the new code just has to be careful to remove users from accessor (aka trustee) orgs before removing or deleting them in the owning org.

Another major piece of cleanup/overhaul had to do with enumerating users in accessor organizations.  Since the conceptual framework for accessors vs. owners is that the owner org owns the domains that the users are in, whereas the accessors are just granting entitlements to the users in those domains, the whole structure assumes that all users must be in the owner org and only some of them are in the accessor org.  This means that, when we go to enumerate accessors, we want to ignore any dashboard users we find that weren't in the owner org.  In addition, we never want to update attributes in the accessor orgs, because trustees are not allowed to do that.  Finally, when we find an unmatched user that we want to operate on, we need to keep track which accessors we've actually seen that user in, so that we don't just go trying to remove every user from every accessor.  (We could do this, of course, but it's overkill, and currently the server gives errors if you try to delete a user who is not actually in the org.)

The work done here included (in no particular order):

* move to umapi_client 2.1, because it has compatibility fixes for changes in the wire protocol around removing users from organizations and deleting them.
* take advantage of the umapi_client upgrade to correctly delete user accounts.
* update the format of the unmatched users list, so that we also track the org of the unmatched user as well as the user key.  This allows being precise about which users need to be removed from which orgs.
* rename "nonexistent" users to be "unmatched" users in all public contexts.  @phil-levy may or may not like this change.
* rename "orphan users" (the internal word in the code for unmatched) to be "strays".
* make sure that the default new account type is inherited and used by the ldap and csv connectors.
* fix the stats to make sense in all cases
* rework the way we track excluded and included users so it works for accessors as well as the owner org.
* rework the way we track unmatched users by keeping a global map from organizations to the user keys that were deleted in those organizations.  This works even when you haven't loaded the directory or dashboard, as when you are removing users on an input list.
* changed input delete list handling so we don't bother loading directory or dashboard users.

The code now seems dead stable and works with multiple organizations quite well.  The one problem is a strange umapi crash on an unexpected server response when you do --remove-entitlements-for-non-existent-users on multiple orgs.  The crash happens *after* all the server-side processing has been done, so essentially the run completes successfully and then umapi_client immediately throws a ServerError.  I'll have to figure this out for the GM build, but for now testing can proceed on RC2.