fix #142 - logging of excluded users.
This was trickier than it appeared!  Now that the logging of matched, excluded, and updated users is all at the debug level, we don't need to be so careful about logging them.  Basically, we should always log them unless there is not chance they might be affected by their status.