(No issue) Fix issues related to code inspection of all files.
This started as a long-needed clean-up of things like redundant parentheses and incorrect None comparisons, but it discovered some subtle and important inconsistencies and edge-case bugs, so it was well worthwhile.  There was only one place where any actual refactoring had to be done in order to ensure consistency between different areas in the LDAP connector code; everything else was basically cosmetic plus discovery of some variable name collisions that altered semantics in unexpected ways.