Fix #198: secure RSA keys via passphrase in keystore
Since Windows doesn't support long keystore values, we can't put the RSA private keys into the keystore.  So we have added support for encryted RSA private keys (either in files or data), and we have added a credential value `priv_key_pass` (in the clear or in the secure store) which is the passphrase to decrypt the private key.

Note that encrypted RSA private keys can be in one of two formats: PKCS#5 or PKCS#8.  (See [this article](https://github.com/kjur/jsrsasign/wiki/Tutorial-for-PKCS5-and-PKCS8-PEM-private-key-formats-differences) for examples of what they look like.)  PKCS#8 keys cannot be brute-force attacked the way that PKCS#5 keys can be, so they are a lot more secure (and allow for much simpler passphrases).  Unfortunately, our version of PyCrypto doesn't support PKCS#8 private keys, and there is no released version that does (only an alpha version).  So we are sticking with PKCS#5 format for now.