Update to: fix #198: secure RSA keys via passphrase in keystore
As of the last commit, we only supported PKCS#5 encrypted private keys.  But with this update, we transition from the old pycrypto module to the brand new (and available on Windows from PyPI) pycryptodome module, which has plug-compatible functionality but runs on modern Pythons *and* has full PCKS#8 support!  So now we support all forms of encrypted private keys!