Fix #182: Allow indirect membership in directory groups.
This changes the LDAP connector functionality to actually query for the members of each group, rather than trying to deduce them from the group membership attributes.

As corollary, this also fixes #129, because we no longer need to read the entire directory in order to make sure we've seen every member of every group.  Since we are querying for the groups directly, we don't have to read the whole directly unless we are directed to.