fix #236: allow users to be pushed rather than sync'd to Adobe.
Adds a `--strategy` command line argument (default 'sync', other value 'push') that controls whether users are fetched from Adobe and sync'd or not.  In the push case, the users are just blindly created against the Adobe org(s) and (if `--process-groups` is specified) any mapped groups are added to them.  This will be idempotent for users that exist on the Adobe side, and will not remove any mapped groups (only add them).