fix to fixes of #257 and #258: issues revealed by testing
The fix for #257 was broken from the get-go, referring to the wrong local variable. This is why one should test before pushing a commit...

The fix for #258 was broken in the case of exceptions that have unicode messages (py2-only).  The py2 exception __str__ method doesn't handle the presence of unicode exception messages, but logging can handle unicode.  So rather than invoking the exception's str handler, we explicitly use logging format on the exception message itself.

The fix for #258 was then applied everywhere we format an exception message, in case any of them have unicode strings.