Cleanup of creation logic for Adobe users.
In order to fix #293, the logic around creating Adobe users had to be revamped.  The original logic was that all Adobe users would be created in the primary umapi, and when they were created there they would also be created in any secondaries for which they had group mappings.  But this fails when secondary group mappings are added *after* the user is created in the primary, as often happens for customers who add Stock consoles.  So now things work as follows:
* New Adobe users are always created in the primary umapi, which is expected be the org that owns all directories where users are created.
* New Adobe users are only created in a secondary umapi if they have groups there.

Revamping the logic around creates allowed streamlining and merging the (hitherto separate) code paths for push and sync:
* Push and Sync are now identical except for how they choose which Adobe users are new (push says they are all new, sync knows which are new).
* All creates in the primary umapi update attributes if requested (because the primary should own the directories); all attributes in secondaries don't update attributes (because the secondaries should trust the directories).

As part of the revamp, I found a spot where Adobe ID attributes were not being updated.  This was cleaned up in order to fix #286 again.

Finally, the order of group operations when users were created had been remove then add, while when a user was being updated it was add then remove.  Since there's some small, theoretical chance that product provisioning of configurations that contain two groups might be affected by the order of group operations, both sequences have been updated to use the remove then add ordering.  There's no issue for this, because we have no actual evidence that it matters.