Fix #294: keep statistics on user changes in secondaries.
The tricky part of this was to separate out which users are being added to the primary *and* the secondary, as opposed to which users are being updated by being added to the secondary.

Getting this right meant renaming a lot of the statistics to more clearly say what they are.